Changing readme, adding authentication file (for initialization)
